OK.... AudioALSA now uses the simple snd_pcm_set_param call to fill in the hw and sw params in the "normal" way.
There appears to be a problem with the interaction between ALSA and
pulse audio that is caused by recent changes in Fedora?

Now we're at a point where the N200 wakes up and receives
very briefly then goes silent.  A RX to TX to RX transition
restores the audio.

So this looks like a problem in the TX/RX state machine and getting
the RX restarted.   That would show up in the output audio file...
More work tomorrow.