Lots of progress.
Now have network control (via port 49172) for TR and band switching
through an N200.  Setup the rest of the infrastructure for the B200
and others.

Alas, something got broken in the TR switching so that in this version
the transmitter LO or something like it is left on after the first
transition from TX to RX mode.  I think this has something to do with
an abandoned experiment in B210 10GHz tuning, but I can't find the
cause.