Fixed integer-N support for UBX.
In particular, this now uses the proposed feature for int-n tuning
on the UBX that allows one to select the step size (R divisor).
Careful choice of frequency steps means that the LO is parked outside
of the band of interest for all tuning settings except for multiples
of 25MHz.