Fixes for audio gain (now the gain control does reasonable things, and is far less sensitive -- no more blown out speakers....)
Fixes for intN tuning, especially for UBX to chose the ref step
to avoid multiples of the reference falling in band....  Work in
progress, but it has cleaned up a few spurs in the 2m band.

Includes an initial, but abandoned, attempt to separate tuning widget
from the USRPCtrl.   A bit too ambitious for right now.