Bug report in https://sourceforge.net/p/sodaradio/discussion/general/thread/8165f370/
indicated that the Band Configuration menu didn't work at all (displayed
things like "SAVE" and "OPEN" and such) on Ubuntu 14.04

User was building against wxWidgets 3.0 which was more particular about
using programmer chosen numbers as wxIDs.. This was a bit of sloppy programming
in the band menu configuration.  It has been fixed.