For some versions of the USRP platform + daughtercard configurations, the antenna names are not "TX/RX" "RX2" and "CAL".  I don't own any of those, but they exist.
So, to accommodate such hardware (and also, in anticipation of someday
supporting Lime platforms where the antennas have all kinds of names.
Names like "Who" and "What" and "I Don't Know") the TX and RX antenna
choice lists now get loaded from hints sent to the GUI from the radio.

This required the addition of the RX_ANT_NAME, and TX_ANT_NAME messages.

This set also adds the RX/TX GAIN_RANGE messages for a change that is
soon to come: actual range indications on the gain sliders. (Though
the audio slider will still go to 11.)