1. cleaned up CMakeLists file to remove some cliplead messages.
2. Added specifier for hamlib port to startup.  Not clear if that will
be useful, but users who launch multiple radios from the same host may
find it so.

3. renamed qtgui/main.cpp to qtgui/main_sodaradio.cpp

4. Added checking to test for the presence of an active SoDaServer when
there should not be one.  This uses the "is there a file present" test
in the ~/.local/share/SoDaRadio directory.

5. Moved the unix domain socket files to the ~/.local/share/SoDaRadio directory

6. improved audio gain on FM demods in BaseBandRX.cxx to make the audio volume
similar to the volume from the CW and SSB demods for "respectable" signals.

7. Added "lockfile" support in SoDaServer.cxx -- see comment number 4.

This is getting close to release time.

The radio does not yet support transverter mode.  Need to look at what I did
in the wx version to make this work.

Next: probably need to remove wx widgets part of the kit, as I don't
intend to support it or fix bugs.  wx is dead.