Made enough changes that the version is now 5.0.2
1. Removed all traces of the original wxgui.  Directory is gone,
and refs in the CMakeLists.txt file are gone.

2. Cleaned up CMakeLists.txt file (removed old junk, fixed an ordering
problem, made it marginally more readable.)

3. GuiParams.cxx needs to signal when "--help" appears on the command
line.  In this case, we don't start the application.  That's more
complicated than it looks, as QApplication needs to be told to quit
and then we must return from main.

4. Removed Demod_test and TSIP_Test from the exp (test unit) builds.
The Demod code was an abandoned attempt at a generic demodulator
class -- not worth the effort.  The TSIP code was a really bad implementation
of a parser for Trimble GPS report streams.  This has been replaced with
the interface to the gpsd daemon.

5. Removed Demod... and TSIP... code from src/SoDaServer build.

6. Fixed a bug where the last used Mode was over-ridden on startup for
reasons that I still can't fully explain.