The log handler was writing empty log files.  That's a bit of an annoyance, so the closing code now checks to see if the log file is 0 length.  If the current output log file is empty, it will be deleted on exit.
If a new log file is opened, then the old file is closed.  A zero
length file will be deleted.

While testing, I found a memory leak that was triggered by calls
to libgps.  After much experimentation, and the addition of a
switch to the soda server command:

SoDaRadio --serverargs "--report_file soda_mem_use.dat"

we have it fixed.

1. Added ProcInfo.[ch]xx to SoDaServer with a test (ProcInfo_test.cxx)
in the exp directory.

2. Added a switch to SoDaServer to record memory usage from /proc/self/statm

3. Added support for --serverargs to the SoDaRadio gui command line.

4. In GPSmon.cxx -- changed "if(gps_waiting" to "while(gps_waiting" to
ensure we don't fall behind the link to the gps server.

5. Added "waitForJoin" method to SoDa::Thread -- this is a join
with a timeout.  Useful for a thousand and one things around the
house, but especially the watchdog/memory check in SoDaServer.cxx

--- This is ready for 5.0.3 rpm builds....