Ubuntu 14.04 (LTS) uses an older version of the gnu C++ compiler that predates the change to ofstream::open in C++11.  Originally, the file name parameter for ofstream::open was char*, but the new standard also provides for a string& parameter.
This was long overdue, but it is incompatible with pre C++11 compiler/libraries.
So, src/ProcInfo.cxx has been modified to "do the right thing".

Thanks to a diligent user who reported the problem.

As this is a change that effects buildability on a currently supported
platform, the version number has been bumped to V5.0.4.