Fixed hamlib handler that did not handle strings of commands from cqrlog
CQRlog sends commands like "fmv" meaning, give me the frequency, mode,
and vfo selection.

Alas, the original hamlib interface (in soda_hamlib_handler) assumed
that there was but one command per line.  This was a bad assumption.
Though the hamlib guys are pretty smart.  There are no single letter
commands 'g' 's' or '\' -- so if we see one of those at the start of a
line, we know that we're going to process a command string.  Otherwise
we're going to process the string one character at a time.

I think this will work.  Preliminary test against wsjtx worked.