Added "Carrier" button toggle event to GUI.  This enables "dead carrier" mode for the transmitter while in CW_U or CW_L modes.  (This feature is used as an aid in tuning up, and also as a marker/beacon for very weak signal microwave stuff.)
Fixed the "dead carrier" tests in USRPTX.cxx

Removed unused beacon envelope from CWTX.[ch]xx

Bumped the version to 5.2.0 as the beacon feature is not just a bugfix.

Added notes (Notes.txt) from spectrum analyzer measurements.  Typical
clean power output levels are 10dBm or so.  There were a few
surprises, but nothing to cause much concern.

BaseBandRX now calculates dB power in the outgoing audio buffer. This
is not yet used for anything, but will eventually be coupled to a
query that may arrive from the GUI. (to implement an audio output
meter)

Improved error messages in soda_listener so that a message emitted
when a message send (to the GUI) fails, it will print the name of
the failing function.

Removed status (noisy debug) messages in hamlib listener and handler.

Added WSPR filter mode with passband of 1400 to 1600 Hz
for WSJTx WSPR (propagation beacon) mode. Filters are in
BaseBandRX.cxx.  The filter names are enumerated in Command.hxx
and reported to the gui in USRPCtrl.cxx.

Added external clock reference button to the SETUP tab.  This uses
the 10MHz ref input to drive the oscillator chains.

Audio restart with ALSA was really really broken.  There was a problem
with restarting ALSA audio output after a buffer under-run.  This was
extremely intermittent before version 5.2 -- like once in a hundred
T/R cycles.  But after the improvements in resampler latency and
oscillator cost, the race occurred much more frequently.  A few more
tweaks got it to happen for at least half of all T/R cycles.

ALSA is as broken as it ever was.  Stream management is either
completely beyond my comprehension or it is broken.  Either way,
SoDaRadio can't rely on it, so now the BaseBandRX module <never>
turns off the audio output stream.  In TX mode it stuffs the output
channel with empty buffers.  Further, USRPRX does not turn off the
RX stream from the radio.  This gives us a continuous IF stream and
audio stream, so they never lose synchronization.

These changes were significant enough to warrant a bump in the
SoDa revision to 5.3.0