Fixed allocation of hardware parameter block for ALSA objects.
Until this rev the setupParams method allocated the hardware parameters
block on the stack.  This is a really bad idea as the parameter block got
used <after> the setupParams call returned.  This error came about because
I did a naive cut-and-paste of some cookbook code without understanding
each line.

Sigh.  It is amazing that it took 4+ years for this to appear.

The bug was found after numerous instances of the audio going dead after
a few hours (or even minutes) of operation with wsjtx.  This error had not
appeared prior to Fedora 28, but it was waiting in the weeds all this time.

The code is being tested now.  If it runs for two days without the audio
problem appearing, then I am inclined to declare victory, but here is no
proof that this was or is the proximate cause for the audio failures.  (It
is simply the most glaring issue.)