Bumped the version to 6.0.0 as this is a big change
Moving toward using Qt for all audio functions.

Currently, RX audio is through Qt and the GUI while TX audio is
still in the SoDaServer process.

The RX audio needs some testing, as early experiments with WWV 10Mhz
were less than impressive.  However, the distortion is likely to have
been QRN rather than bad software -- raw carrier from the Fluke 6080
sounded very clean.

The RX did manage to allow WSJTx to decode some FT8 signals on 40m
despite the RX overload.  So audio appears to be sufficiently good for
a modem.

Audio on the 162.4 WX radio is passable. With the 6kHz audio filter,
there was a bit of noise on the signal -- it was not quite broadcast
quality.  But there does not appear to be a difference in quality
with respect to the ALSA/PA RX audio implementation.

Major changes:
src/AudioQt.[ch]xx
New:  Based on AudioALSA, but have removed all local audio handling.
Instead AudioQt creates a server socket.  When BaseBandRX calls "send"
on the rx audio unit, the bits are sent over the socket to a listener
in the GUI.

src/BaseBandRX.[ch]xx
Removed buffering and buffer management and flow/rate management.
This simplifies the audio chain, but makes the recovery from clock
differences a little crude. (Rate matching is up to the gui soda_audio_listener
process.)

qtgui/soda_audio_listener.[ch]pp
New: connects to the audio server socket, and copies arriving bits to a
QtAudioOutput object.  This is harder than it looks.

src/UDSockets.[ch]xx
added a param to "put" that sends the buffer
without a message length prefix.

src/SoDaServer.cxx
Select the AudioQt object. (as opposed to the slowly rotting AudioALSA object.)

qtgui/all the rest
Changes to add controls and connections for the audio device. (Including
device selection!)