More work on rate matching radio stream to audio output. The previous attempt saw the audio fall behind by several seconds over the course of a day.  It is clear that activity on the host computer can cause the audio system to fall behind by as much as one second over a very short (5 second) window.
The new scheme limits the run-ahead and also stuffs zeros until we have
at least 5 frames worth of audio available in the circular buffer.

I suspect we'll need to implement sample cutting to allow the audio
to catch up, but sample cutting doesn't help if the fall-behind events
get one second behind.

This may be a linux real-time issue.