In the continuing effort to improve the RX audio handling, the audio listener is now in its own Qt Thread object.
The audio listens on a socket for incoming audio
packets from the radio, writes them via a callback to the
audio output. When the socket and the audio output were both
being served from the main event loop, some GUI events could
cause a sudden delay in the socket or audio handler that caused
the audio to fall up to 1.5 seconds behind (!).  If this happened
a few times things could get really out of hand.

There is now a status label in the lower left corner of the GUI
that shows how many seconds of audio are in the circular buffer.
This amounts to a guess at how far behind the audio output is from
the RX receive chain.

In this version, if the audio falls more than 1 second behind, the
audio listener will flush the circular buffer and "catch up" with
a possible unpleasant hiccup in the audio out.

If the circular buffer has less than about 0.25 seconds stored,
the listener will supply an empty output buffer until the
circular buffer has enough samples in it to keep "ahead" of the
audio device.

This version appears to be ready for prime time.