Holy cow!  Recording works!
Poke the record button and it writes a WAV file whose name is the
current date/time.

For reasons I don't understand, the file dialog that sets the directory
is splatting a GTK warning that lots of folks have reported.  Alas, like
the weather, nobody has done anything about it.