Fixed doxygen complaints.
Removed a debug print or two...

Bumped the version to 6.4 because....

The Wx version of SoDaRadio followed the cursor in the
panadapter/waterfall and retuned on each click.  That wasn't a good
choice for weak signal detection where knowing that the RX spurs won't
move is useful.  However, having the display scroll as the radio is
tuned would be nice.  So, the UI should answer a MB3 button click in
the display area by changing the frequency and centering the spectrum
around it.  MB3 now does the same thing as MB1 -- it sets the RX
frequency -- and it also re-centers the display (waterfall, or
spectrogram) on the chosen frequency.  (MB3 is like poking MB1 and
then "Center RX Freq.")

Next:
During testing with B2xx, it became clear that the lack of a squelch for NBFM
mode is somewhat irritating.  As an SDR, the radio has lots of knowledge about
the quality of the incoming signal at the output of the FM IF.  At minimum we
can look for a minimum IF input signal level.