NBFM has a squelch setting on the SETTINGS tab.  It is unitless, and its setting is relative to the RF gain setting, so the two interact... sigh. The alternative would scan the whole IF span and look for the in-band energy to be greater than "squelch dB" above the average density.  But that isn't very convenient with the way the IF strip (see src/BaseBandRX.cxx) is organized.
In operation, once a particular squelch level is set, the RF gain can be
adjusted up/down to get a little more control over the breakpoint.

The squelch setting is stuck on the Settings page because there isn't
sufficient room on the main panel for a control that is only used in NBFM
mode.

Hang time for the squelch is about 1/4 second or so.  This might be a bit
long for some, but it is a start.