There were problems with SoDaRadio on Fedora 30 -- the audio listener code in the GUI enumerated many audio devices, some of which could not support the 48kS/s sample rate.  (SURROUND20, for instance).  If there was a "saved" choice of audio output device stored in the saved settings file, it may select one of these devices, causing the audio listener startup to fail.
The changes here addressed this issue:
1. At enumeration time, only those audio devices that support the sample
rate and format requirements will be listed in the combobox audio output
device menu.
2. If the stored selection from the settings file is invalid, the first
item in the audio output device list (the default device) will be selected.

In addition, the gpsd people changed the api for gps_read.  There is now
conditional code in GPSmon.cxx to select the right version of gps_read
based on the API revision.

This commit also fixes a problem with gcc 9.x where it complained about four
poorly written lambda expressions in SoDaThreadRegistry.cxx.