Move dependencies on gpsd from GPSmon to a separate header/implementation.
The gpsd client library changes the API in compile-breaking ways from
one version to the next.  API changes are expected on major version
changes for a library, but most responsible library maintainers
keep backward compatability for a little while.  Only sociopaths
would do otherwise.

So "fix_gpsd_ugliness.[cxx,hxx]" separate out the gpsd code.  Handling
gpsd calls requires lots of "#if this_condition" and "#if that_condition"
which makes the code uglier than it should be.  There is enough
ugliness in the world without spreading more.  So, the ugliness is
now sequestered in a sty of its own.

Unfortunately, libgpsd is the only game in town for GPS libraries. It
would be simple to build support for one or two NMEA devices, but the
proliferation of "special cases" and odd-ball TSIP protocol from
Trimble make development of a library depend on lots of users.  libgps
has sucked the air out of the room. So we're stuck with it.

The libgps folks warn us about the fact that they don't preserve any
kind of backward compatability with a (paraphrasing here) "check the
major version number and make sure it is one you've seen" and (again
paraphrasing) "if you don't we'll tank your build."  No deprecation
warnings. No support for older code.

So now fix_gpsd_ugliness.cxx checks for the API version.  This isn't
available at CMake time, unfortunately, so a user might think we're
compiling in GPS support when we don't. The fix_... code prints a message
at compile time if the gps library was found, but has an API version
that is newer than we expect.

ADDED:
src/fix_gpsd_ugliness.[cxx,hxx]
All gpsd code is here now.

src/GPSmon.[cxx,hxx]
Now refers to a GPSD object and retrieves time, lat, and lon independent
of the choices made in libgps.

src/CmakeLists.txt
1) Added new fix_gpsd_ugliness stuff
2) Removed a bunch of .hxx files from the SRCs list as cmake figures
that out on its own.

Commit comments
A bucket of bile brought up by continuing frustration with the libgps team.