Working on removing boost dependencies.
First step is to remove boost::format.  That's what this commit does.

boost::format is replaced with SoDa::Format -- a more intelligent formatter
that doesn't carry quite as much baggage or flexibility as boost::format.

The SoDa::Format dependency is satisfied either by an installed version
of SoDa::Format, or by building the format library within the SoDaRadio
build as an ExternalProject.  In the case of MacOS, the SoDa::Format
library *must* be installed, as the private build will run into problems
with MacOS lack of support for static libraries.