Removed all boost dependencies. (LibUHD still depends on boost, sigh.)
Updated qt components to fit with Qt5.

But this version has a problem with delay between an action in the gui
(startup, button press, etc) and the actual action taking effect. There
is some problem with the interaction between Qt and ???

Qt also complained of requestActivate being unsupported in Wayland. This
is a known bug that the Qt folks don't appear to want to fix.
The workaround is to define an environment variable:
$ export QT_QPA_PLATFORM=xcb

This is unfortunate.