All I wanted to do was to replace the older resampler and OSfilter widgets with the new ones from SoDaLibs.
Unfortunately, the newer ones are pretty good about discipline. So they've got
smart pointers instead of naked pointers, and vectors instead of float * and such.

But that broke the mailbox stuff.  Fortunately, SoDaLibs contains a
new Mailbox scheme that is cleaner than the original SoDa implementation.

But that broke some more stuff.

It's broken turtles all the way down.