The port to qt6 has not gone well.
In fact Fedora43 was a disaster.  I just couldn't get even the most basic
Qt6 audio stuff to work.  So I'm switching to Ubuntu in the hopes that
things are better.  Sadly support for tools and libraries on Fedora is
starting to wither - most of the online hints are for Ubuntu.